backend/requirements.txt
# --- DATA & MATH ---
yfinance==1.2.0
pandas==3.0.1
pandas-ta==0.4.71b0
numpy==2.2.6
numba==0.61.2
llvmlite==0.44.0
multitasking==0.0.12

# --- DATABASE & CACHE ---
SQLAlchemy==2.0.48
psycopg2-binary==2.9.11
redis==7.3.0
peewee==4.0.2

# --- API, WEB & NETWORKING ---
fastapi==0.135.2
uvicorn==0.42.0
pydantic==2.12.5
pydantic_core==2.41.5
starlette==1.0.0
httpx==0.28.1
httpcore==1.0.9
requests==2.32.5
urllib3==2.6.3
websockets==16.0
anyio==4.12.1
h11==0.16.0
idna==3.11
sniffio==1.3.1
jiter==0.13.0
curl_cffi==0.13.0

# --- AI & ORCHESTRATION ---
anthropic==0.86.0
python-dotenv==1.2.2
beautifulsoup4==4.14.3
soupsieve==2.8.3
tqdm==4.67.3
langchain==0.3.19
langchain-openai==0.3.7
langgraph==0.2.74

# --- UTILS & CORE DEPENDENCIES ---
click==8.3.1
python-dateutil==2.9.0.post0
typing_extensions==4.15.0
typing-inspection==0.4.2
annotated-doc==0.0.4
annotated-types==0.7.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
distro==1.9.0
docstring_parser==0.17.0
frozendict==2.4.7
platformdirs==4.9.4
protobuf==7.34.1
pycparser==3.0
pytz==2026.1.post1
six==1.17.0
openai==1.61.1